+++ dla/iija/_data_utils.py
@@ -39,9 +39,9 @@ def read_data_all():
     proj = proj.dropna(how='all') 
     proj['summary_recipient_defined_text_field_1_value'] = proj['summary_recipient_defined_text_field_1_value'].fillna(value='None')
     
-    new_codes = to_snakecase(pd.read_excel(f"{GCS_FILE_PATH}/FY21-22ProgramCodesAsOf5-25-2022.v2.xlsx"))
-    code_map = dict(new_codes[['iija_program_code', 'new_description']].values)
-    proj['program_code_description'] = proj.program_code.map(code_map)
+    # new_codes = to_snakecase(pd.read_excel(f"{GCS_FILE_PATH}/FY21-22ProgramCodesAsOf5-25-2022.v2.xlsx"))
+    # code_map = dict(new_codes[['iija_program_code', 'new_description']].values)
+    # proj['program_code_description'] = proj.program_code.map(code_map)
     
     return proj
 
@@ -62,6 +62,31 @@ def change_col_to_integer(df, col):
     df[col] = df[col].str.split(' ').str[-1]
     
     return df
+
+
+#for use in the following function identify_agency
+def add_name_from_locode(df, df_locode_extract_col):
+    #read in locode sheet
+    locodes = to_snakecase(pd.read_excel(f"gs://calitp-analytics-data/data-analyses/dla/e-76Obligated/locodes_updated7122021.xlsx"))
+    
+    #extract locode, make sure it i has the same spots
+    df['locode'] = df[df_locode_extract_col].apply(lambda x: x[1:5])
+    #make sure locode column is numeric
+    df['locode'] = pd.to_numeric(df['locode'], errors='coerce')
+    
+    #merge
+    df_all = (pd.merge(df, locodes, left_on='locode', right_on='agency_locode', how='left'))
+    
+    df_all = df_all.rename(columns={'agency_name':'implementing_agency',
+                                   'locode':'implementing_agency_locode'})
+    
+    #if we use other locode list then drop these columns
+    df_all.drop(columns =['active_e76s______7_12_2021_', 'mpo_locode_fads', 'agency_locode'], axis=1, inplace=True)
+        
+    return df_all
+
+
+
 
 # def tokenize(texts):
 #     return [nltk.tokenize.word_tokenize(t) for t in texts]
